split out identity and key for auth
