+++ auth/context.go
@@ -9,6 +9,7 @@ import (
 	"github.com/goadesign/goa"
 	"github.com/goadesign/goa/middleware/security/jwt"
 	"github.com/pkg/errors"
+	"github.com/zenoss/zenkit/claims"
 	"github.com/zenoss/zenkit/logging"
 )
 
@@ -49,5 +50,25 @@ func DevModeMiddleware(h goa.Handler) goa.Handler {
 			req.Header.Set("Authorization", devJWT)
 		}
 		return h(ctx, rw, req)
+	}
+}
+
+func JWTValidatorFunc(m JWTValidator) goa.Middleware {
+	return func(h goa.Handler) goa.Handler {
+		return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
+			if err := m(ctx); err != nil {
+				return errors.WithStack(err)
+			}
+			token := jwt.ContextJWT(ctx)
+			stdClaims := claims.StandardClaimsMap(token.Claims.(jwtgo.MapClaims))
+			ident := &tokenIdentity{stdClaims}
+			ctx = WithIdentity(ctx, ident)
+			ctx = goa.WithLogContext(ctx, "user_id", ident.ID())
+			if len(ident.Tenant()) == 0 {
+				message := "Unable to retrieve tenant from token"
+				return errors.WithStack(errors.New(message))
+			}
+			return h(ctx, rw, req)
+		}
 	}
 }
